general layout
general layout, still working on button appearance & header
